Code hygiene moved helpers out of constants
